+++ Dziedziczenie/kwadrat.h
@@ -9,17 +9,21 @@ public:
 	Kwadrat();
 	Kwadrat(int a, int b);
 	~Kwadrat();
-	void SetX(int x);
-	void SetY(int y);
-	int GetX();
-	int GetY();
+	void SetX(double x);
+	void SetY(double y);
+	double GetX();
+	double GetY();
 	double ObliczPole();
 	double ObliczObwod();
 	void WypiszDane();
 
+	// Hermetyzacja - zmienne dostêpne tylko dla funkcji klasy kwadrat.
+	// Klasa szescian która dziedziczy po kwadracie mo¿e mieæ do nich dostêp tylko za pomoc¹ funkcji klasy kwadrat
+
+	//friend class Szescian; //deklaracja przyjaŸni, klasa Szeœcian mia³a by bezpoœredni dostêp do x i y z klasy Kwadrat
 private:
-	int x;
-	int y;
+	double x;  
+	double y;
 };
 
 #endif // !KWADRAT_H
